requirements.txt
pyinstaller
openai
python-dotenv
httpx
Pillow
websockets
tiktoken
